(WIP) Dynapcnn Visualizer initial implementation.
The API has changed as  does not have  function. Rather  has method  which takes in a  object. This is done to avoid bloating of code in dynapcnn_network.py (As Sadique predicted, if it had been the functional programming option, it would mean the connection part should take place in DynapcnnNetwork, which is a lot of verbose code in connecting graphs. Furthermore there would have been circular function calls where both objects have to call methods from one another. ). The only missing part now is the communication part with graphs.